// lib/api/generated/models/models-newsletter-subscription-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Huerray Backend
 * Comprehensive API for connecting content creators with brands for marketing campaigns and gig opportunities
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@thrypes.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface ModelsNewsletterSubscriptionResponse {
    'created_at'?: string;
    'email'?: string;
    'first_name'?: string;
    'id'?: string;
    'last_name'?: string;
    'status'?: string;
    'subscribed_at'?: string;
    'unsubscribed_at'?: string;
    'updated_at'?: string;
}